ReAct and Multiagent/requirements.txt
crewai
streamlit
python-dotenv
pydantic
crewai[tools]
